Make the HTTP async with asyncio.
The good news is now we are async (though I've maintained the public API being synchronous for now).

The bad news is that this drops python2, emits some warnings which I'm failing to diagnose, and looking at the asyncio internals, appears to monkeypatch some global state so god only knows if this seemingly stateless code is really thread-safe.